% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_funnel.R
\name{benchmark_funnel}
\alias{benchmark_funnel}
\title{Benchmark funnel plot boundaries}
\usage{
benchmark_funnel(exp, scvtx = 3, alphax = 0.95)
}
\arguments{
\item{exp}{expected values from indirect standardization (observed/expected)}

\item{scvtx}{benchmark SCV , by default 3}

\item{alphax}{alpha level, by default 95\%}
}
\value{
exp

upperCI Upper limit O/E of benchmark Poisson-Gamma distribution

lowerCI Lower limit O/E of benchmark Poisson-Gamma distribution
}
\description{
This function produces data for drawing funnel plot boundaries that correspond to the 95\% confidence intervals
of a Poisson-gamma distribution corresponding to a desired systematic component of variation, the benchmark SCV.
}
\examples{

data(nc.sids, package= "spData")

obs<-nc.sids$SID74

exp<-nc.sids$BIR74*sum(nc.sids$SID74)/sum(nc.sids$BIR74)

outlier<-outlier_stats(obs,exp,scvtx=3, alphax=.95)

funnel<-benchmark_funnel(exp,scvtx=3, alphax=.95)

library(ggplot2)
ggplot(funnel,aes(exp,upperCI))+
geom_line()+
geom_line(aes(y=lowerCI))+
geom_point(data=outlier$outlier_stats,aes(x=exp, y=obs/exp,color=is_outlier))+
xlab("Expected cases")+
ylab("Observed/Expected")+
labs(color="Outlier")+
theme_minimal()

}
