% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats functions.R
\name{.pg_var}
\alias{.pg_var}
\title{Internal functions}
\usage{
.pg_var(obsx, expx)
}
\arguments{
\item{obsx}{observed}

\item{expx}{expected}

\item{scvtx}{benchmark scv}

\item{alphax}{threshold level}
}
\description{
Internal functions
}
\keyword{internal}
