% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_stats.R
\name{outlier_stats}
\alias{outlier_stats}
\title{Outlier statistics and outlier gains}
\usage{
outlier_stats(obsx, expx, scvtx = 3, alphax = 0.95)
}
\arguments{
\item{obsx}{observed cases for a given population}

\item{expx}{expected cases from standardization of given population}

\item{scvtx}{benchmark SCV, default 3}

\item{alphax}{alpha level, default .95}
}
\value{
obs

exp

upper_lim_qnbinom Upper limit O/E of benchmark Poisson-Gamma distribution

lower_lim_qnbinom Lower limit O/E of benchmark Poisson-Gamma distribution

is_outlier Identifes O/E as upper or lower outlier (else is NULL)

pval_ref Probability of O/E ratio in benchmark Poisson-Gamma distribution

outlier_diff Difference between outlier and benchmark upper or lower limit

upper_outlier_gain System gains under scenario of normalizing upper outliers

lower_outlier_gain System gains under scenario of normalizing lower outliers
}
\description{
Outlier detection from  benchmark Poisson-Gamma distribution (a distribution drawn from a benchmark SCV).
Outlier gains are the additions or reductions (depending on whether it is a desired outcome or not) to the system that would occur
under a scenario where all identified upper outliers or lower outliers are brought into the 95\% bounds of the benchmark distribution
}
\examples{

data(nc.sids, package= "spData")

obs<-nc.sids$SID74

exp<-nc.sids$BIR74*sum(nc.sids$SID74)/sum(nc.sids$BIR74)

outlier_stats(obs,exp,scvtx=3, alphax=.95)

}
