% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_stats.R
\name{scv_stats}
\alias{scv_stats}
\title{Poisson-Gamma systematic component of variation}
\usage{
scv_stats(obsx, expx)
}
\arguments{
\item{obsx}{observed cases for a given population}

\item{expx}{expected cases from standardization of given population}
}
\value{
obs

 exp

pval Probability of O/E ratio in empirical Poisson-Gamma distribution

smoothed Empirical bayes smoothing of O/E ratio

scv Systematic component of variation

scv_var Bootstrapped variance of the scv

scv_CI_low Bootstrapped lower 95\% confidence interval of SCV

scv_CI_upp Bootstrapped upper 95\% confidence interval of SCV
}
\description{
Extension of the empbaysmooth function in DCluster. Adds a computation of the SCV as 100* \emph{nu}/\emph{alpha^2}.
Computes from simulations a bootstrapped variance and 95\% CI of the SCV.
}
\examples{

data(nc.sids, package= "spData")

obs<-nc.sids$SID74

exp<-nc.sids$BIR74*sum(nc.sids$SID74)/sum(nc.sids$BIR74)

scv_stats(obs,exp)


}
