% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_funnel.R
\name{benchmark_funnel}
\alias{benchmark_funnel}
\title{Funnel plot boundaries for a Poisson-Gamma distribution with a benchmark SCV}
\usage{
benchmark_funnel(exp, scvtx = 3, alphax = 0.95)
}
\arguments{
\item{exp}{expected values from indirect standardization (observed/expected)}

\item{scvtx}{benchmark SCV , by default 3}

\item{alphax}{alpha level, by default 95\%}
}
\value{
exp

upperCI upper confidence intervals

lowerCI lower confindence interval
}
\description{
Funnel plot boundaries for a Poisson-Gamma distribution with a benchmark SCV
}
\examples{

data(nc.sids, package= "spData")

exp<-nc.sids$BIR74*sum(nc.sids$SID74)/sum(nc.sids$BIR74)

benchmark_funnel(exp,scvtx=3, alphax=.95)

}
