% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_stats.R
\name{outlier_stats}
\alias{outlier_stats}
\title{Outlier detection from empirical and benchmark Poisson-Gamma distribution}
\usage{
outlier_stats(obsx, expx, scvtx = 3, alphax = 0.95)
}
\arguments{
\item{obsx}{observed}

\item{expx}{expected}

\item{scvtx}{benchmark SCV, default 3}

\item{alphax}{alpha level, default .95}
}
\value{
obs

exp

upper_lim_qnbinom upper limit O/E of benchmark Poisson-Gamma distribution

lower_lim_qnbinom lower limit O/E of benchmark Poisson-Gamma distribution

is_outlier identifes O/E as upper or lower outlier (else is NULL)

pval_ref probability of O/E ratio in benchmark Poisson-Gamma distribution

outlier_diff difference between outlier and benchmark upper or lower limit

upper_outlier_gain system gains under from scenario of normalizing upper outliers

lower_outlier_gain system gains under from scenario of normalizing lower outliers
}
\description{
Outlier detection from empirical and benchmark Poisson-Gamma distribution
}
\examples{

data(nc.sids, package= "spData")

obs<-nc.sids$SID74

exp<-nc.sids$BIR74*sum(nc.sids$SID74)/sum(nc.sids$BIR74)

outlier_stats(obs,exp,scvtx=3, alphax=.95)

}
