% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv_stats.R
\name{scv_stats}
\alias{scv_stats}
\title{Poisson-Gamma systematic component of variation}
\usage{
scv_stats(obsx, expx)
}
\arguments{
\item{obsx}{observed}

\item{expx}{expected}
}
\value{
obs

 exp

pval probability of O/E ratio in empirical Poisson-Gamma distribution

smoothed empirical bayes smoothing of O/E ratio

scv systematic component of variation

scv_var bootstrapped variance of the scv

scv_CI_low bootstrapped lower 95\% confidence interval of SCV

scv_CI_upp bootstrapped upper 95\% confidence interval of SCV
}
\description{
Poisson-Gamma systematic component of variation
}
\examples{

data(nc.sids, package= "spData")

obs<-nc.sids$SID74

exp<-nc.sids$BIR74*sum(nc.sids$SID74)/sum(nc.sids$BIR74)

scv_stats(obs,exp)


}
